Attach yaml to md

Attach yaml to md

attach_yaml2md(yaml, md, out = md)

Arguments

yaml
A list or a string. If a list is passed to yaml, it is treated as a yaml structure as returned by yaml::yaml.load(). If a file path is passed, yaml::read_yaml() will be used to read the file.

md
Path to the .md or .Rmd file.

out
Path to the output .md file. Defaults to the path specified in md, which means overwriting the input file.